split mpi function from newtoniterationutilities
