+++ opm/autodiff/NewtonIterationUtilities.hpp
@@ -21,7 +21,6 @@
 #define OPM_NEWTONITERATIONUTILITIES_HEADER_INCLUDED
 
 #include <opm/autodiff/AutoDiffBlock.hpp>
-#include <boost/any.hpp>
 #include <vector>
 
 namespace Opm
@@ -58,10 +57,6 @@ namespace Opm
                             const std::vector< AutoDiffBlock<double> >& eqs,
                             Eigen::SparseMatrix<double, Eigen::RowMajor>& A,
                             AutoDiffBlock<double>::V& b);
-
-    /// Return true if this is a serial run, or rank zero on an MPI run.
-    bool isIORank(const boost::any& parallel_info);
-
 } // namespace Opm
 
 #endif // OPM_NEWTONITERATIONUTILITIES_HEADER_INCLUDED
